Set filterOutStylingRules method in utils
