+++ app/angular/src/server/angular-cli_utils.js
@@ -23,6 +23,24 @@ function getAssetsParts(resolvedAssetPath, assetPath) {
   };
 }
 
+function isStylingRule(rule) {
+  const { test } = rule;
+
+  if (!test) {
+    return false;
+  }
+
+  if (!(test instanceof RegExp)) {
+    return false;
+  }
+
+  return test.test('.css') || test.test('.scss') || test.test('.sass');
+}
+
+export function filterOutStylingRules(config) {
+  return config.module.rules.filter(rule => !isStylingRule(rule));
+}
+
 export function isBuildAngularInstalled() {
   try {
     require.resolve('@angular-devkit/build-angular');
